WorkingBCIFan_to_Emotiv/requirements.txt
websocket-client==1.6.4
pyserial==3.5